requirements.txt
mysql_connector_repackaged==0.3.1
numpy==1.23.5
Pillow==10.3.0
python_bcrypt==0.3.2
scikit_learn==1.4.1.post1
streamlit==1.33.0
tensorflow==2.12.0
tensorflow_intel==2.12.0
streamlit==1.32.2


